KOGITO-8680 - Move authentication handling to resource class
