+++ quarkus/addons/process-svg/runtime/src/main/java/org/kie/kogito/svg/dataindex/QuarkusDataIndexClient.java
@@ -20,7 +20,6 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
@@ -31,8 +30,7 @@ import org.kie.kogito.svg.ProcessSVGException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.quarkus.security.credential.TokenCredential;
-import io.quarkus.security.identity.SecurityIdentity;
+import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -47,17 +45,15 @@ import static java.util.Objects.nonNull;
 public class QuarkusDataIndexClient implements DataIndexClient {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(QuarkusDataIndexClient.class);
-    private SecurityIdentity identity;
+
     private Vertx vertx;
     private WebClient client;
     private String dataIndexHttpURL;
 
     @Inject
     public QuarkusDataIndexClient(@ConfigProperty(name = "kogito.dataindex.http.url", defaultValue = "http://localhost:8180") String dataIndexHttpURL,
-            SecurityIdentity identity,
             Vertx vertx) {
         this.dataIndexHttpURL = dataIndexHttpURL;
-        this.identity = identity;
         this.vertx = vertx;
     }
 
@@ -78,16 +74,19 @@ public class QuarkusDataIndexClient implements DataIndexClient {
     @Override
     public List<NodeInstance> getNodeInstancesFromProcessInstance(String processInstanceId, String authHeader) {
         String query = getNodeInstancesQuery(processInstanceId);
-        CompletableFuture<List<NodeInstance>> cf = new CompletableFuture<>();
-        client.post("/graphql").putHeader("Authorization", getAuthHeader(authHeader)).sendJson(JsonObject.mapFrom(singletonMap("query", query)), result -> {
-            if (result.succeeded()) {
-                cf.complete(getNodeInstancesFromResponse(result.result().bodyAsJsonObject()));
-            } else {
-                cf.completeExceptionally(result.cause());
-            }
-        });
+        Future<List<NodeInstance>> future = client.post("/graphql")
+                .putHeader("Authorization", authHeader)
+                .putHeader("content-type", "application/json")
+                .sendJson(JsonObject.mapFrom(singletonMap("query", query)))
+                .map(response -> {
+                    if (response.statusCode() == 200) {
+                        return getNodeInstancesFromResponse(response.bodyAsJsonObject());
+                    } else {
+                        return null;
+                    }
+                });
         try {
-            return cf.get();
+            return future.toCompletionStage().toCompletableFuture().get();
         } catch (Exception e) {
             throw new ProcessSVGException("Exception while trying to get data from Data Index service", e);
         }
@@ -108,10 +107,4 @@ public class QuarkusDataIndexClient implements DataIndexClient {
         }
     }
 
-    protected String getAuthHeader(String authHeader) {
-        if (identity != null && identity.getCredential(TokenCredential.class) != null) {
-            return "Bearer " + identity.getCredential(TokenCredential.class).getToken();
-        }
-        return authHeader;
-    }
 }
