+++ quarkus/addons/process-svg/runtime/src/test/java/org/kie/kogito/svg/dataindex/QuarkusDataIndexClientTest.java
@@ -21,8 +21,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import io.quarkus.security.credential.TokenCredential;
-import io.quarkus.security.identity.SecurityIdentity;
 import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -32,8 +30,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.lenient;
-import static org.mockito.Mockito.mock;
 
 public class QuarkusDataIndexClientTest {
 
@@ -58,7 +54,7 @@ public class QuarkusDataIndexClientTest {
             "  }\n" +
             "}";
 
-    QuarkusDataIndexClient client = new QuarkusDataIndexClient(null, null, null);
+    QuarkusDataIndexClient client = new QuarkusDataIndexClient(null, null);
 
     @Test
     public void testGetNodeInstancesFromResponse() {
@@ -83,7 +79,7 @@ public class QuarkusDataIndexClientTest {
 
     @Test
     public void testSetupMalformedURL() {
-        QuarkusDataIndexClient testClient = new QuarkusDataIndexClient("malformedURL", null, null);
+        QuarkusDataIndexClient testClient = new QuarkusDataIndexClient("malformedURL", null);
         assertThrows(MalformedURLException.class, () -> testClient.setup());
     }
 
@@ -114,18 +110,4 @@ public class QuarkusDataIndexClientTest {
         assertTrue(webClientOptions.isSsl());
     }
 
-    @Test
-    public void testGetTokenWithSecurityIdentity() {
-        String token = "testToken";
-        TokenCredential tokenCredential = new TokenCredential(token, "Bearer");
-        SecurityIdentity identity = mock(SecurityIdentity.class);
-        lenient().when(identity.getCredential(TokenCredential.class)).thenReturn(tokenCredential);
-        QuarkusDataIndexClient testClient = new QuarkusDataIndexClient(null, identity, null);
-        assertThat(testClient.getAuthHeader("")).isEqualTo("Bearer " + token);
-    }
-
-    @Test
-    public void testGetTokenWithoutSecurityIdentity() {
-        assertThat(client.getAuthHeader("")).isEmpty();
-    }
 }
